Fix: NoteService

+++ src/main/java/goit/ua/mynotespet/notes/service/NoteService.java
@@ -4,6 +4,7 @@ import goit.ua.mynotespet.notes.dto.request.CreateNoteRequest;
 import goit.ua.mynotespet.notes.dto.request.UpdateNoteRequest;
 import goit.ua.mynotespet.notes.dto.response.NoteResponse;
 import goit.ua.mynotespet.notes.entity.Note;
+import goit.ua.mynotespet.notes.exception.NoteNotFoundException;
 import goit.ua.mynotespet.notes.repository.NoteRepository;
 import goit.ua.mynotespet.users.entity.User;
 import goit.ua.mynotespet.users.service.UserService;
@@ -33,10 +34,12 @@ public class NoteService {
         return mapToResponse(savedNote);
     }
 
-    public NoteResponse getNoteById(Long id) {
+    public NoteResponse getNoteById(Long id, String username) {
         Note note = noteRepository.findById(id)
-                .orElseThrow(() -> new IllegalArgumentException("Note not found! with id: " + id));
-
+                .orElseThrow(() -> new NoteNotFoundException("Note not found! with id: " + id));
+        if (!note.getUser().getUsername().equals(username)) {
+            throw new NoteNotFoundException("Note not found!");
+        }
         return mapToResponse(note);
     }
 
@@ -48,18 +51,30 @@ public class NoteService {
     }
 
     @Transactional
-    public NoteResponse updateNote(UpdateNoteRequest request, Long id) {
+    public NoteResponse updateNote(Long id, UpdateNoteRequest request, String username) {
         Note note = noteRepository.findById(id)
-                .orElseThrow(() -> new IllegalArgumentException("Note not found! with id: " + id));
+                .orElseThrow(() -> new NoteNotFoundException("Note not found! with id: " + id));
 
-        note.setTitle(request.getTitle());
-        note.setContent(request.getContent());
+        if (!note.getUser().getUsername().equals(username)) {
+            throw new NoteNotFoundException("Note not found!");
+        }
+        if (request.getTitle() != null && !request.getTitle().equals(note.getTitle())) {
+            note.setTitle(request.getTitle());
+        }
+        if (request.getContent() != null && !request.getContent().equals(note.getContent())) {
+            note.setContent(request.getContent());
+        }
+
         Note updateNote = noteRepository.save(note);
         return mapToResponse(updateNote);
+
     }
 
-    public void deleteNoteById(Long id) {
-        Note note = noteRepository.findById(id).orElseThrow(() -> new IllegalArgumentException("Note not found! with id: " + id));
+    public void deleteNoteById(Long id, String username) {
+        Note note = noteRepository.findById(id).orElseThrow(() -> new NoteNotFoundException("Note not found! with id: " + id));
+        if (!note.getUser().getUsername().equals(username)) {
+            throw new NoteNotFoundException("Note not found!");
+        }
         noteRepository.delete(note);
     }
